posts and tags list components
We are looping through all of models and creating a show component for
each.

We are yielding inside show component, this way whatever passed as body
of this component will be repeated inside each show component.

We are yielding with an argument, caller of this component can use this
data to create their HTML.